Add C++ programs for creating classes and objects
This commit adds three C++ programs that demonstrate the creation of classes and objects. The first program creates a class called \MyClass\ and an object called \myObj\. The second program creates a class called \Car\ and two objects called \carObj1\ and \carObj2\. The third program also creates a class called \Car\ but includes a constructor to initialize the object properties.